comp_pract/lr_6/requirements.txt
Flask
flask-cors
flask-socketio
pillow